:lipstick: refactor : styled-component 분리
