+++ src/components/deleteModal/DeleteModal.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { useParams } from "react-router";
 import axios from "axios";
-import { BasicBtn, CancelBtn } from "../../elements/buttons/button";
+import { BasicBtn, CancelBtn } from "../../elements/buttons/Button";
 import {
   BtnWrap,
   ModalBg,
