+++ src/components/deleteModal/style.tsx
